first

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -19,7 +19,8 @@ if __name__ == '__main__':
 
     states = session.query(State).filter(
         State.name.like('%a%')).order_by(State.id).all()
-    for state in states:
-        session.delete(state)
+    if states:
+        for state in states:
+            session.delete(state)
         session.commit()
     session.close()
